refactor: 검색어 강조 기능 모듈화
